modifyAllProject
